requirements.txt
together==0.3.0
flask==2.3.3
flask-cors==3.0.10
python-dotenv==1.0.0
gunicorn==21.2.0
pyjwt==2.8.0
pymongo==4.6.0
